FAC-963:impl:Move subjectinsight packages to insight package
