+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/subjectinsight/CreateSubjectInsightRequestDto.java
@@ -1,0 +1,4 @@
+package org.flickit.assessment.core.adapter.in.rest.insight.subjectinsight;
+
+public record CreateSubjectInsightRequestDto(String insight) {
+}
