+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/subjectinsight/CreateSubjectInsightRestController.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.core.adapter.in.rest.subjectinsight;
+package org.flickit.assessment.core.adapter.in.rest.insight.subjectinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.subjectinsight.CreateSubjectInsightUseCase;
-import org.flickit.assessment.core.application.port.in.subjectinsight.CreateSubjectInsightUseCase.Param;
+import org.flickit.assessment.core.application.port.in.insight.subjectinsight.CreateSubjectInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.subjectinsight.CreateSubjectInsightUseCase.Param;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
