+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/insight/subjectinsight/InitSubjectInsightRestController.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.core.adapter.in.rest.subjectinsight;
+package org.flickit.assessment.core.adapter.in.rest.insight.subjectinsight;
 
 import lombok.RequiredArgsConstructor;
 import org.flickit.assessment.common.config.jwt.UserContext;
-import org.flickit.assessment.core.application.port.in.subjectinsight.InitSubjectInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.subjectinsight.InitSubjectInsightUseCase;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,7 +19,7 @@ public class InitSubjectInsightRestController {
     private final UserContext userContext;
 
     @PostMapping("/assessments/{assessmentId}/subjects/{subjectId}/init-insight")
-    public ResponseEntity<Void> initSubjectInsight(@PathVariable("assessmentId")UUID assessmentId,
+    public ResponseEntity<Void> initSubjectInsight(@PathVariable("assessmentId") UUID assessmentId,
                                                    @PathVariable("subjectId") Long subjectId) {
         var currentUserId = userContext.getUser().id();
         useCase.initSubjectInsight(toParam(assessmentId, subjectId, currentUserId));
