--- flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/subjectinsight/CreateSubjectInsightRequestDto.java
@@ -1,4 +1,0 @@
-package org.flickit.assessment.core.adapter.in.rest.subjectinsight;
-
-public record CreateSubjectInsightRequestDto(String insight) {
-}
