+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/insight/subjectinsight/GetSubjectInsightUseCase.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.subjectinsight;
+package org.flickit.assessment.core.application.port.in.insight.subjectinsight;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
