+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/subjectinsight/ApproveSubjectInsightServiceTest.java
@@ -1,8 +1,8 @@
-package org.flickit.assessment.core.application.service.subjectinsight;
+package org.flickit.assessment.core.application.service.insight.subjectinsight;
 
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
-import org.flickit.assessment.core.application.port.in.subjectinsight.ApproveSubjectInsightUseCase;
+import org.flickit.assessment.core.application.port.in.insight.subjectinsight.ApproveSubjectInsightUseCase;
 import org.flickit.assessment.core.application.port.out.subjectinsight.ApproveSubjectInsightPort;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
